+++ pages/setting/indieblog.less
@@ -1,0 +1,66 @@
+.tag-title {
+  font-weight: bold;
+  font-size: 20px;
+  margin: 0 0 10px;
+  display: inline;
+}
+
+.list {
+  margin: 20px 0 0 0;
+  display: flex;
+  align-items: center;
+
+  .list-content {
+    flex: 8;
+  }
+
+  .list-title {
+    flex: 2;
+    align-self: center;
+  }
+}
+
+.title {
+  font-size: 16px;
+  font-weight: 400;
+  color: #333;
+  line-height: 28px;
+  margin-right: 20px;
+}
+
+.center {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.w-10px {
+  width: 10px;
+}
+
+.site {
+  &-available {
+    flex: 1;
+    display: flex;
+  }
+
+  &-link {
+    flex: 2;
+    color: #542DE0;
+    font-weight: 600;
+    &:visited {
+      color: #542DE0;
+    }
+  }
+}
+
+.link {
+  color: #542DE0;
+  &:visited {
+    color: #542DE0;
+  }
+}
+
+.no-create-title {
+  margin-bottom: 10px;
+}
